Avoid int vs size_t warning in OpenSSL code

+++ src/lib/prov/openssl/openssl_ec.cpp
@@ -74,7 +74,7 @@ int OpenSSL_EC_curve_builtin(int nid)
       return -1;
       }
 
-   for(int i=0; i < num; ++i)
+   for(size_t i = 0; i < num; ++i)
       {
       if(builtin_curves[i].nid == nid)
          {
